#12 abstracted the app bar and side bar to be used in other pages in future
